26: Renamed classes for cryptographic operation and renamed its package
